# Optimize tensor operation