--- a/src/bgg/sampler.rs
+++ b/src/bgg/sampler.rs
@@ -4,6 +4,7 @@ use crate::poly::polynomial::Poly;
 use crate::poly::sampler::{DistType, PolyUniformSampler};
 use crate::poly::PolyMatrix;
 use crate::poly::{params::PolyParams, sampler::PolyHashSampler};
+use crate::utils::print_memory_usage;
 use itertools::Itertools;
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;
@@ -130,10 +131,18 @@ where
 
         let gadget = S::M::gadget_matrix(params, 2);
         let encoded_polys_vec = S::M::from_poly_vec_row(params, plaintexts.to_vec());
-        let second_term = encoded_polys_vec.tensor(&(secret_vec.clone() * gadget));
+        let rhs = secret_vec.clone() * gadget;
+        println!(
+            "Tensor product input sizes: encoded_polys_vec={:?}, rhs={:?}",
+            encoded_polys_vec.size(),
+            rhs.size()
+        );
+
+        let second_term = encoded_polys_vec.tensor(&rhs);
+        print_memory_usage("Before tensor product");
 
         let all_vector = first_term - second_term + error;
-
+        print_memory_usage("After tensor product");
         parallel_iter!(plaintexts)
             .enumerate()
             .map(|(idx, plaintext)| {