Core idea: I have a `S` matrix of size 2x12, `other` matrix of size 3x3 and identity `I` matrix of size 4x4. Goal is to perform `S * (I ⊗ other)`.

Current implementation: First perform `partial = (I ⊗ other)` and then `out = S * partial`.

Optimized implementation: Directly compute `out = S1*other | S2*other | S3*other | S4*other` where each `Si` is the slice of `S` of size 2 x 3

Results:

Reduced peak memory consumption from 4.78 GB to 353 MB

Improved performance

Before
```
Time to obfuscate: 139.643509375s
Time for evaluation: 102.822680333s
Total time: 242.466189708s
```

After
```
Time to obfuscate: 105.465769375s
Time for evaluation: 86.621996s
Total time: 192.087765375s
```